Refactor node server by moving db models and dbsync script into server.
* Add env file into server in prep for breakout into microservice.
