+++ server/index.js
@@ -5,7 +5,7 @@ const graphqlHTTP = require("express-graphql");
 const { buildSchema } = require("graphql");
 const cors = require("cors");
 const request = require("request");
-const models = require("../models/index");
+const models = require("./models/index");
 const prettyPrintResponse = require("./helpers");
 const client = require("./plaid");
 const appRoutes = require("./routes/pocketgoblin.routes.js");
